updater: Use smaller blocksize for clearing partition table
Also run sync to ensure the content is surely written down to disk
this addresses a pesky issue where sometimes new partitions did not
come out good and junk content is seen in /boot partition and it is
mounted as read-only and all files are not available. Sometimes these
files could be dtbo files which means certain feaures may not work
even after system is booted ok.

Signed-off-by: Khem Raj <raj.khem@gmail.com>