linux-ti-staging: Add USB HID to kernel proper
When connecting USB keyboard/mouse to beagelbone, it does
not work unless we install kernel-module-usbhid which is also pulled as
dependency of kernel-modules meta package. Unlike other SBCs like rpi,
bbb does not add kernel-modules to image via MACHINE_EXTRA_RRECOMMENDS
so as a result, keyboard and mouse do not work out of box.

This option is adding the needed kernel module to zImage itself which
then should work well.

Signed-off-by: Khem Raj <raj.khem@gmail.com>