yoe-installer-image: Make it a true SD installer
This image was always intended to be an installer image which is written
to a SD card and can then format and install it into EMMC or other
permanent media that is final on the system. Therefore this image can
act as factory install image.

Here we decouple it from other images and make it a super target which
will include .upd file corresponding to FACTORY_INSTALL_IMAGE set in
site.conf which can be project spefic, a fallback is yoe-simple-image

For coral-dev we use coral-rfs9000-image

Now we can flash coral-dev_Coral-RFS9000-v0.9.903.wic.xz to SD card
put the board in SD boot mode and let it install eMMC from this SD card
once it boots all the way, pop SD card out and switch back to eMMC boot
mode and reboot the system to now fully boot from eMMC

Signed-off-by: Khem Raj <raj.khem@gmail.com>