yoe: Add few more handy template include files
commerical-eula is useful for world builds but should not be used for
product builds without review.

report-error is good for reporting build errors to yocto community error
server or customized to internal repository if you setup one.

scanbuild is useful to run static analysis on either full distro or a
single recipe

Signed-off-by: Khem Raj <raj.khem@gmail.com>