updater: Only overlay based on "tmpfs" or "persistent"
Current behavior is to fallback to tmpfs overlay if project does not
define it. Change this to such that unless "persistent" or "tmpfs" is
specified we do not apply an overlay at all. Which means it makes rootfs
r/w

Define OVERLAY_TYPE=none in updater platform file to disable overlay
completely.

Signed-off-by: Khem Raj <raj.khem@gmail.com>