envsetup: Build bmap-tools-native explicitly
yoe_check_install_dependencies computes dependencies for flashing image
and image flasher uses bmaps to speed up writing SD cards. It really
makes a lot of difference especially when images are large in size

build-sysroots target has become leaner and droppped over dependencies
which is a good thing in general but it means that now it is  not
building bmap-tools that we need. So be explicit about it before
building sysroots, so it gets into the staged sysroot

Signed-off-by: Khem Raj <raj.khem@gmail.com>